Fleshed out more of the WebOS features. Added in the codebase for WebOSTV.net to speed up this part of the development. There may be a way to do this with just using the nuget package, but I couldn't figure it out.
Have an open issues set up to figure out what the current state of this project is. Not sure if it would make more sense for me to keep on rolling my own version, update his version, or take over the project. Hopefully he will reply and we'll be able to hash it out.